requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.0.0
gunicorn==20.1.0
idna==3.2
psycopg2-binary==2.9.1
python-dotenv==0.19.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
